# fix(rbac): add/remove roles in bulk to increase performance