--- a/workspaces/rbac/plugins/rbac-backend/src/audit-log/audit-logger.ts
+++ b/workspaces/rbac/plugins/rbac-backend/src/audit-log/audit-logger.ts
@@ -62,6 +62,11 @@ export type RoleAuditInfo = {
   members: string[];
 };
 
+export type RolesAuditInfo = {
+  groupingPolicies: string[][];
+  source: Source;
+};
+
 export type PermissionAuditInfo = {
   policies: string[][];
   source: Source;