## Hey, I just made a Pull Request!

Fixes two issues that were leading to slow start ups of the RBAC plugin. 

The first was that we were not adding and removing roles in bulk. Prior to this PR we would bulk add/remove the created permission policies. However, we would individually add/remove the roles. This was due to needing to log the creation/deletion of the role whenever it occurs. My suggested fix is to bulk add/remove and log in bulk similar to how we do it with the permission policies. Information can still be obtained about through the use of the role metadata from the UI and API. 

The second was that we were removing all of the roles and permissions at start up regardless of whether they are meant to persist. This seemed like an oversight from a previous change and I went ahead and adjusted the behavior back to what it originally was (only add/remove if there is a need). This was a relatively small change with a big impact so I went ahead and included it into this PR.

For comparison, originally with 1000 roles (based on the example given in issue #2552) it would take 22 seconds on a first time start up with subsequent restarts averaging 63 seconds. Now, a first start will take roughly 2 seconds with subsequent restarts averaging 1.5 seconds. Additionally, no unnecessary logs for updates are occurring that only make a change to the updated time field.

### Fixes
- Fixes: #2552 
- Fixes: [RHIDP-6180](https://issues.redhat.com/browse/RHIDP-6180)

#### :heavy_check_mark: Checklist

<!--- Please include the following in your Pull Request when applicable: -->

- [x] A changeset describing the change and affected packages. ([more info](https://github.com/backstage/community-plugins/blob/master/CONTRIBUTING.md#creating-changesets))
- [ ] Added or updated documentation
- [ ] Tests for new functionality and regression tests for bug fixes
- [ ] Screenshots attached (for UI changes)
- [x] All your commits have a `Signed-off-by` line in the message. ([more info](https://github.com/backstage/community-plugins/blob/master/CONTRIBUTING.md#developer-certificate-of-origin))